so I had the signatures all messed up, also fixes so that double is forced on someone

+++ RationalNumber.java
@@ -19,7 +19,7 @@ public class RationalNumber extends Number
   }
 
   public double getValue(){
-    double inital = Math.round(numerator*10000/denominator);
+    double inital = Math.round(numerator*10000/(double)denominator);
     return inital/10000;
   }
